extract inner classes
